src/online_stores/requirements.txt
redis==7.1.0
pandas>=2.3.3
pyarrow>=22.0.0
fastapi==0.127.0
uvicorn==0.40.0
python-dotenv>=1.0.0